+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/commands.py
@@ -4,7 +4,6 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core.commands import CliCommandType
-from azure.cli.core.util import empty_on_404
 
 from ._format import (
     registry_output_format,
@@ -80,7 +79,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('list', 'acr_list')
         g.command('create', 'acr_create')
         g.command('delete', 'acr_delete')
-        g.command('show', 'acr_show', exception_handler=empty_on_404)
+        g.show_command('show', 'acr_show')
         g.command('login', 'acr_login', table_transformer=None)
         g.command('show-usage', 'acr_show_usage', table_transformer=usage_output_format)
         g.generic_update_command('update',
@@ -95,13 +94,15 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('import', 'acr_import')
 
     with self.command_group('acr credential', acr_cred_util) as g:
-        g.command('show', 'acr_credential_show', exception_handler=empty_on_404)
+        g.show_command('show', 'acr_credential_show')
         g.command('renew', 'acr_credential_renew')
 
     with self.command_group('acr repository', acr_repo_util) as g:
         g.command('list', 'acr_repository_list')
         g.command('show-tags', 'acr_repository_show_tags')
         g.command('show-manifests', 'acr_repository_show_manifests')
+        g.command('show', 'acr_repository_show')
+        g.command('update', 'acr_repository_update')
         g.command('delete', 'acr_repository_delete')
         g.command('untag', 'acr_repository_untag')
 
@@ -109,7 +110,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('list', 'acr_webhook_list')
         g.command('create', 'acr_webhook_create')
         g.command('delete', 'acr_webhook_delete')
-        g.command('show', 'acr_webhook_show')
+        g.show_command('show', 'acr_webhook_show')
         g.command('get-config', 'acr_webhook_get_config', table_transformer=webhook_get_config_output_format)
         g.command('list-events', 'acr_webhook_list_events', table_transformer=webhook_list_events_output_format)
         g.command('ping', 'acr_webhook_ping', table_transformer=webhook_ping_output_format)
@@ -125,7 +126,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('list', 'acr_replication_list')
         g.command('create', 'acr_replication_create')
         g.command('delete', 'acr_replication_delete')
-        g.command('show', 'acr_replication_show')
+        g.show_command('show', 'acr_replication_show')
         g.generic_update_command('update',
                                  getter_name='acr_replication_update_get',
                                  setter_name='acr_replication_update_set',
@@ -139,7 +140,7 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
 
     with self.command_group('acr build-task', acr_build_task_util) as g:
         g.command('create', 'acr_build_task_create')
-        g.command('show', 'acr_build_task_show')
+        g.show_command('show', 'acr_build_task_show')
         g.command('list', 'acr_build_task_list', table_transformer=build_task_output_format)
         g.command('delete', 'acr_build_task_delete')
         g.command('update', 'acr_build_task_update')