+++ src/command_modules/azure-cli-billing/HISTORY.rst
@@ -3,9 +3,9 @@
 Release History
 ===============
 
-0.1.10
-++++++
-* Minor fixes
+0.2.0
++++++
+* BREAKING CHANGE: 'show' commands log error message and fail with exit code of 3 upon a missing resource.
 
 0.1.9
 ++++++